separating components; reply done
