+++ src/Kambaz/Courses/Piazza/ViewPost/FollowupDiscussions/FollowupDiscussions.css
@@ -9,6 +9,7 @@
     border-radius: 2px;
     background-color: #555555;
     width: max-content;
+    margin-left: 7px;
   }
   
   .resolved label,
@@ -54,6 +55,11 @@
     border-top: 1px solid #ccc;
     margin: -5px -7px 0 0;
   }
+
+  .followup_content_wrapper {
+    border-top: 1px solid #fafafa;
+    background: #fafafa;
+  }
   
   .followup_container .content.container-fluid {
     padding-left: 12px;
@@ -67,6 +73,7 @@
   .avatar {
     margin-top: 1px;
     margin-right: 4px;
+    margin-left: 7px;
     border-radius: 0;
     width: 34px;
     height: 34px;
